g4zdc updates
I ran the zdc classes through clang and made some minor changes.

cppcheck picked up some things that would be best by changing. There were several issues of ```EICG4ZDCHitTree.cc:122:6: error: Non-local variable 'layerType' will use pointer to local variable 'v_layerType'. [danglingLifetime]```. For the nTuple, we filled several branches with vectors. In this class we declared two sets of vectors and copied the values across. Only the inital vector was cleared, not the one that fills the nTuple. Some of the branches were hit locations so its possible to accidentally add previous hits to the next event:

Assume the first set of hits is 30 hits. This is fine and we size the vector appropriately. The next set of hits is 40 hits, again no problem. The third set of hits is 35 hits, we now have kept elements 35-39 from the previous hit set and biased the nTuple

To-do: run this class in our simulation